Use `ProductMonikerRid` instead of `CoreSetupRid` for SDK RID.
This commit changes `NETCoreSdkRuntimeIdentifier` to be based off of
`ProductMonikerRid` instead of `CoreSetupRid`.

`CoreSetupRid` is set to the host RID for the build, by default.  For most
architectures, this was fine.  However, for ARM/ARM64 builds, we cross-build,
so `CoreSetupRid` is actually x64 and not ARM/ARM64.

As a result, the default apphost being used isn't built for ARM so `dotnet run`
fails with an exec format error when trying to exec the apphost.

Instead, we should be using the RID of the product being built rather than the
host RID.

Fixes dotnet/cli#10291.