Address PR feedback
* AfterTargets="Build" -> BeforeTargets="AfterBuild"
* Ensure SdkResolver.csproj is built before
* Fix `dotnet pack` of redist.csproj without having built everything first

Also make SdkResolver build layout on all architectures and only pack on
x86 condition. The reason for only packing on one architecture is to prevent
competition between build legs for publishing the same bits. x86 is chosen
because we use this flavor of the sdk in corext and it's easier if all the
corext packages can be obtained from the same build gesture/drop.