Remove rhel 7
We should remove this leg. Centos coverage is enough according to dotnet/dotnet-buildtools-prereqs-docker#194

We happen to use rhel 7 not only in test but actually creating the rpm packages. So we need to move the base image to create rpm from rhel 7 to centos and that also requrie an update to centos since centos 7 is no longer supported. I get the latest centos image in https://github.com/dotnet/versions/blob/master/build-info/docker/image-info.dotnet-dotnet-buildtools-prereqs-docker-master.json