[release/5.0.1xx-preview8] Port Microsoft.DotNet.Sdk.Internal fixes from master (#8105)
* Revert "Only publish Microsoft.DotNet.Sdk.Internal once (#8016)"

This reverts commit 7e76b75c653b74dd9525bfa7b586a398fa5b4a7d.

* Revert "Publish the SDK sentinel package as a package (#8012)"

This reverts commit a7cf77dd667e4fd560083bc1acbbe02c60c74b89.

* Fix asset manifest file names.
Set the asset manifest OS and change the name of the installers manifests.
Currently arcade defaults to using OS-Platform as the asset manifest file name.
Installer does a ton of building in docker containers, which ends up generating
many overlapping names. This is a symptom and also likely hiding some non-determinsm.

To fix, change the name of the blob asset manifest name to include a "-installers" suffix.
Then alter the property names around a bit so that AssetManifestOS is set
prior to publish.proj's publish target being run.