Make CLISDK MSI set the InstallLocation when bundle does not
We changed the bundle to not pass in default values for DOTNETHOME
because it cannot calculate them in all cases: there is no way to
conditionally set a path which does not exist.  As a result, the MSI
needs to calculate them so that it can write the appropriate defaults.