Fix upgrade of Arm64 SDK
The previous 5.0 Arm64 will have written a bad value for InstallLocation.

We tried to ignore this case on non-Arm64, since we know those don't matter.

We were previously assuming folks would uninstall all previous SDKs that
wrote the bad values on Arm64.  Now that we're supporting the 5.0 Arm64
SDK we can't do that, since we need it to be able to upgrade the old 5.0
Arm64 SDK.