[release/6.0.4xx] Fix cross-gen for DumpMinitool
Exclude DumpMinitool from cross gen because it prevents it from starting.

On Windows when we create a hang dump of a process, we need to ensure that the native minidump api is called from a process that has the same bitness as the target process. If we fail to do that dotnet test produces a 64-bit dump of 32-bit process which is unusable for debugging. For this reason, there is DumpMinitool.*.exe shipped for each supported architecture, which must remain as is rather than being crossgened.

Fix for dotnet/sdk#25421
Fix for #13942