Override some versions for nonshipping runtime packages. (#15045)
Source-build 7.0.0 created these packages with an incorrect version number which is stuckthe previously-source-built PVP for the 7.0.1 release.  This causes a problem if bootstapping because the Microsoft versions of these packages don't actually exist, so trying to download them fails.  This change allows us to override the version read out of PVP and set it to the actual version used by runtime.