PVP flow support Ported changes from work done in the VMR.
This change supports the ability to have PVP flow for source build. This means that repo builds only receive the versions of packages they actually have dependencies on. This setting may be enabled per repo in the repo projects via setting PackageVersionPropsFlowType to 'DependenciesOnly'.

To support this, we need to make a few changes:
- The previously source-built packages must be written per repo
- The current source-built packages must be written per repo.
- The build input props task must be updated to support the two flow types.

Package version props creation has been changed to write a per-repo props file in all cases, even if there is no filtering. This required a bit of refactoring overall. However, when De0endenciesOnly is not enabled, build behavior not changed.