Update MAUI baseline manifests for 8 Preview 4
In c29cfc12, we updated the MAUI baseline manifests so our public preview 2 packages were in the .NET SDK for 8 Preview 3.

Update to MAUI .NET 8 Preview 3 manifests for Preview 4.

We should also be able to remove the `xamarin-impl` feed from the `NuGet.config`, as that was only used for early .NET 6 releases.