Only list supported RIDs for ILCompiler (#16545)
Attempting to publish AOT on an unsupported host, or with an
unsupported target RID, will produce an error in the SDK before trying
to restore the ILCompiler packages. This uses the RID list in
`KnownILCompilerPack` as the source of truth for what is considered
supported. Then with improved error reporting
(https://github.com/dotnet/sdk/pull/32943), this will give enough info
to the user about why a scenario isn't supported for aot.

Failing before restore does mean that advanced users who want to test
with unsupported ILC packages need to use a different
workaround. Instead of adding a `PackageReference`, they'll need
something like this:

```xml
<ItemGroup>
  <KnownILCompilerPack Update="@(KnownILCompilerPack)">
    <ILCompilerRuntimeIdentifiers>unsupported-rid</ILCompilerRuntimeIdentifiers>
  </KnownILCompilerPack>
</ItemGroup>
```