Test suite refactoring
There are assumptions in the test infrastructure that all tests are
based around testing the SDK output. But that's not always the case.
This refactors the test suite to have a new TestBase class which can be
used for tests that don't target the SDK and SmokeTest class has been
renamed to SdkTests for those tests which do test the SDK. Also factored
out the logic for parsing exclusion files into the Utilities class.