Use the VMR commit hash for dotnet --info
This defines a new environment variable `DotNetGitCommitHash` which
contains the commit hash of the VMR. Individual repositories can use this
to find the VMR's commit hash and then embed/use it as appropriate.

It then uses this hash when generating the SDK's .version file, so
dotnet --info shows this as the SDK version.

Contributes to https://github.com/dotnet/source-build/issues/3643