Use built-in `SourceRevisionId` instead of `git rev-parse HEAD`
With .NET 8, Source Link is included in the SDK. It defines the
`SourceRevisionId` property for us, and that property already contains
the git commit hash. So lets use that instead of trying to compute it
via `git rev-parse HEAD`.

This should not affect source-build mode of the VMR using a tarball.
source-build writes the minimal git repository metadata [1] to disk, and
that's enough for Source Link to set up `SourceRevisionId` correctly.

[1] https://github.com/dotnet/sourcelink/tree/main/docs#minimal-git-repository-metadata