Generate a props file that can be used to detect MSBuild version mismatches.
We know the minimum and 'bundled' MSbuild versions, but users may build a project with newer SDKs than we expected (specifically when full-framework MSBuild is starting the build of an SDK-style project).
When this occurs, we'd like to automatically condition the use of PackageReferences meant to ensure compatibility of the Roslyn toolchain, so we need to know if we are in this mismatched situation.
A fast and simple way to do this is to 'stamp' the 'expected' version
of MSBuild during product construction and compare that to the
'current' version being used during the actual build.