Show VMR commit for dotnet --info
Remove the logic that overrides the git commit info for each repo. Let
each repo infer the VMR's git commit (via sourcelink).

That will let the installer and sdk repos use the VMR's commit hash and
show that in `dotnet --info`.

This works for both when building the VMR out of git, as well as when
building it from a tarball/archive thanks to sourcelink's support for
minimal git metadata [1].

[1] https://github.com/dotnet/sourcelink/tree/main/docs#minimal-git-repository-metadata

Fixes: dotnet/source-build#3643