Delete SourceBuiltArtifacts after bootstrap
When building the VMR in source-build bootstrap mode, we end up with
duplicate copies of nuget packages - one set in the SourceBuiltArtifacts
tarball, and another in the restored packages directory. We can safely
delete the tarball since it's not longer relevant.

This comes up particularly when building .NET in scenarios where the
actual build is done offline (eg, Fedora). For that scenario, we run the
`./prep-source-build.sh --bootstrap` on a developer machine, then create
a "source archive" with the results of that, and then upload/build that
in the formal offline build systems. In that case, the extra
SourceBuiltArtifacts increases the size of the "source archive" by more
than a gigabyte.