Add top-level controls to VMR for system libraries
Often, when building .NET, we want to build with different
configurations of external libraries. For example, runtime has a copy of
brotli, and we may or may not want to use that.

This introduces top-level properties to the VMR to select whether to use
the bundled copy of a library (eg in, runtime/src/native/external) vs
the system copy of the library.