Handle spaces in properties passed to VMR's build.sh
When building with something like:

    ./build.sh '/p:OfficialBuilder=Red Hat'

The VMR currently fails to build with an error like this:

    MSBUILD : error MSB1008: Only one project can be specified.
    Full command line: 'dotnet/.dotnet/sdk/9.0.100-preview.5.24253.2/MSBuild.dll
          -maxcpucount -verbosity:m -tlp:default=auto dotnet/eng/init-source-only.proj
          /p:DotNetBuildSourceOnly=true /p:OfficialBuilder=Red Hat
          -distributedlogger:Microsoft.DotNet.Tools.MSBuild.MSBuildLogger,dotnet/.dotnet/sdk/9.0.100-preview.5.24253.2/dotnet.dll*Microsoft.DotNet.Tools.MSBuild.MSBuildForwardingLogger,dotnet/.dotnet/sdk/9.0.100-preview.5.24253.2/dotnet.dll'
    Switches appended by response files:
      Switch: Hat

    For switch syntax, type "MSBuild -help"

As the full command line shows, properties with spaces are split into separate
properties. That results in invalid switches to msbuild, which makes msbuild
error out.

Fix this by switching to bash arrays, and use full quoting as needed.