Merged PR 44011: MSBuild Patch for Dependency Upgrades
Patches the fix for https://github.com/dotnet/source-build/issues/4681

By adding the dependencies to the version.details graph and using version props for the dependencies, PVP flow is used and the msbuild packages are updated to the use 6.0.5 and 6.0.2 versions of the runtime packages, respectively.

----
#### AI description  (iteration 1)
#### PR Classification
Dependency update.

#### PR Summary
This pull request updates the commit SHA for the Microsoft.NET.Sdk and Microsoft.DotNet.MSBuildSdkResolver dependencies.
- Updated `Microsoft.NET.Sdk` and `Microsoft.DotNet.MSBuildSdkResolver` SHA in `/eng/Version.Details.xml`